+++ pkg/gateway/service.go
@@ -132,8 +132,10 @@ func newService(bifrost *Bifrost, serviceOptions config.ServiceOptions) (*Servic
 
 	// direct proxy
 	upstreamOptions := config.UpstreamOptions{
-		ID:       uuid.NewString(),
-		Balancer: "round_robin",
+		ID: uuid.NewString(),
+		Balancer: config.BalancerOptions{
+			Type: "round_robin",
+		},
 		Targets: []config.TargetOptions{
 			{
 				Target: hostname,