+++ CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [unreleased]
+
+### Changed
+
+- Change balancer in upstream options and allow to pass params ([#68](https://github.com/nite-coder/bifrost/pull/68) by [0x5487](https://github.com/0x5487))
+
 ## [0.6.0] (2025-08-30)
 
 ### Added